[DISCO-3021] move content_handler folder into utils
