+++ merino/jobs/navigational_suggestions/domain_metadata_extractor.py
@@ -8,7 +8,7 @@ import requests
 from pydantic import BaseModel
 from robobrowser import RoboBrowser
 
-from merino.content_handler.models import Image
+from merino.utils.content_handler.models import Image
 from merino.jobs.navigational_suggestions.utils import (
     REQUEST_HEADERS,
     TIMEOUT,
